movie-recommendation-app/requirements_presentation.txt
python-pptx==0.6.21
Pillow==10.0.0 